quanda/requirements.txt
Authlib~=1.0.0
Django==4.2.4
djongo-fixed==1.3.7
django-cors-headers
gunicorn==21.2.0
psycopg-binary==3.1.10
psycopg==3.1.10
djangorestframework
django-filter
requests ~= 2.31
django-redis==5.3.0
boto3==1.28.30

